git-svn-id: https://zuow-pc:8443/svn/navy@137 47ee70cb-904c-3e43-b3d1-3900189539e7

+++ navy_v3/src/cn/com/hd/navy/importmanage/ExportService.java
@@ -87,6 +87,12 @@ public class ExportService extends BaseService implements IService {
 					projectImg += "#" + img.getImagename();
 				}
 				
+				// 判断源文件是否存在
+				File temp = new File(base+img.getImagepath());
+				if(!(temp.exists())) {
+					continue;
+				}
+				
 				FileUtils.copy(base + img.getImagepath(), targetPath + img.getImagename());
 			}
